+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 0.1.13 - 
 
+- `r_to_py` now supports proper detectiong of r DataFrames and their conversion (indexes and column names are correctly kept)
+- Added `Summary` dataclass that accurately describes the format returned by summary()
+- Extended `summary` to return the whole summary and added `__repr__` for pretty and quick summaries
 
 
 ## 0.1.12 - RDS loading/saving, families functions, default priors functions